+++ client/src/api-calls/progress-updates.js
@@ -48,37 +48,4 @@ const updateProgressUpdate = async ({ therapistMessage, id, options } = {}) => {
   }
 };
 
-const sendFeedback = async ({
-  clearInstructions,
-  problems,
-  clearDemos,
-  noDemos,
-  enjoyableResources,
-  likeMostAndLeast,
-  programmeId,
-  options,
-} = {}) => {
-  try {
-    // const { data } = await axios.post(`${PROGRESS_UPDATES_BASE}`, {
-    //   clearInstructions,
-    //   problems,
-    //   clearDemos,
-    //   noDemos,
-    //   enjoyableResources,
-    //   likeMostAndLeast,
-    //   programmeId,
-    // });
-    // return { data };
-    return true;
-  } catch (error) {
-    const err = handleError(error, options);
-    return { error: err };
-  }
-};
-
-export {
-  getProgressUpdatesById,
-  sendUpdate,
-  updateProgressUpdate,
-  sendFeedback,
-};
+export { getProgressUpdatesById, sendUpdate, updateProgressUpdate };
